Fixed map running bugs for test

+++ include/Entity.hpp
@@ -12,7 +12,7 @@ class Entity : public Drawable
 
 protected:
     float damage; // The viable entities all have these variables
-    sf::Vector2f location;
+    sf::Vector2i location;
     float speed;
     float radius_of_attack;
     int cost;
@@ -25,7 +25,7 @@ public:
 
     //getters
     float getDamage();
-    sf::Vector2f getLocation();
+    sf::Vector2i getLocation();
     float getSpeed();
     float getRadius_of_attack();
     int getCost();
@@ -34,7 +34,7 @@ public:
     //setters
     void setDamage(float new_dmg);
     void setSpeed(float new_sp);
-    void setLocation(sf::Vector2f newloc);
+    void setLocation(sf::Vector2i newloc);
     void setRadius_of_attack(float radius);
     void setCost(int new_c);
     void setisActive(bool active);
